refactor(workflows): extract shared ProposalDiff renderer from ProposalCard (#580)
